update kfptask to be more module friendly
